+++ package.json
@@ -15,11 +15,13 @@
     "check:fix": "biome check --write ."
   },
   "dependencies": {
+    "autoprefixer": "^10.4.21",
     "gsap": "^3.13.0",
     "next": "^15.1.6",
+    "postcss": "^8.5.6",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "tailwindcss": "^4.1.11"
+    "tailwindcss": "^3.4.17"
   },
   "devDependencies": {
     "@biomejs/biome": "^2.1.2",
